Moved all the classes to a separate file
