+++ lib/OpenEVSEDriver.js
@@ -1,0 +1,13 @@
+/* jshint node: true, esversion: 6*/
+
+"use strict";
+
+const EventEmitter = require("events");
+
+//const debug = require("debug")("openevse:OpenEVSEDriver");
+
+module.export = class OpenEVSEDriver extends EventEmitter
+{
+
+};
+
