+++ lib/openevse.js
@@ -6,164 +6,10 @@
 
 "use strict";
 
-const EventEmitter = require("events");
-const simulator = require("./simulator_driver");
-const SerialPort = require("serialport");
 const debug = require("debug")("openevse");
-const Readline = SerialPort.parsers.Readline;
 
-class OpenEVSEError{
-  constructor(type, message = "") {
-    debug("Error "+type+": "+message);
-    this.type = type;
-    this.message = message;
-  }
-}
-
-class OpenEVSERequest
-{
-  constructor()
-  {
-    this._done = () => {};
-    this._error = () => {};
-    this._always = () => {};
-  }
-
-  done(fn) {
-    this._done = fn;
-    return this;
-  }
-
-  error(fn) {
-    this._error = fn;
-    return this;
-  }
-
-  always(fn) {
-    this._always = fn;
-    return this;
-  }
-}
-
-class OpenEVSEDriver extends EventEmitter
-{
-
-}
-
-class OpenEVSEDriverHttp extends OpenEVSEDriver
-{
-  constructor(endpoint)
-  {
-    super();
-    this._endpoint = endpoint;
-    if (endpoint.substring(0, 5) === "https") { this.http = require("https"); }
-    else { this.http = require("http"); }
-  }
-
-  rapi(command, callback = () => {})
-  {
-    var request = new OpenEVSERequest();
-    var url = this._endpoint + "?json=1&rapi="+encodeURI(command);
-    this.http.get(url, (res) => {
-      res.setEncoding("utf8");
-      var body = "";
-
-      res.on("data", (chunk) => {
-        body += chunk;
-      });
-
-      res.on("end", () => {
-        var data;
-        try {
-          data = JSON.parse(body);
-          callback(data.ret);
-          request._always();
-        }
-        catch (e) {
-          request._error(new OpenEVSEError("BadBody", body));
-          request._always();
-        }
-      }).on("error", () => {
-        request._error(new OpenEVSEError("RequestFailed"));
-        request._always();
-      }).setTimeout(6000, () => {
-        request._error(new OpenEVSEError("HTTPTimeout"));
-        request._always();
-      });
-    });
-
-    return request;
-  }
-}
-
-class OpenEVSEDriverSimulator extends OpenEVSEDriver
-{
-  constructor()
-  {
-    super();
-    simulator.onevent((data) => {
-      var eventData = data.split(" ");
-      var event = eventData.shift();
-      this.emit(event, eventData);
-    });
-  }
-
-  rapi(command, callback = () => {})
-  {
-    var request = new OpenEVSERequest();
-    setTimeout(() => {
-      callback(simulator.rapi(command));
-      request._always();
-    }, 1);
-
-    return request;
-  }
-}
-
-class OpenEVSEDriverSerial extends OpenEVSEDriver
-{
-  constructor(endpoint)
-  {
-    super();
-    this.serial = new SerialPort(endpoint, {
-      baudRate: 115200
-    });
-
-    const parser = new Readline({ delimiter: "\r" });
-    this.serial.pipe(parser);
-
-    this.requests = [];
-    parser.on("data", (data) => {
-      if(data.startsWith("$OK") || data.startsWith("$NK"))
-      {
-        if(this.requests.length > 0)
-        {
-          var request = this.requests.pop();
-          request.callback(data);
-          request.request._always();
-        }
-      }
-      else
-      {
-        var eventData = data.split(" ");
-        var event = eventData.shift();
-        this.emit(event, eventData);
-      }
-    });
-  }
-
-  rapi(command, callback = () => {})
-  {
-    var request = new OpenEVSERequest();
-    this.serial.write(command+"\r");
-    this.requests.push({
-      callback: callback,
-      request: request
-    });
-
-    return request;
-  }
-}
+const OpenEVSEError = require("OpenEVSEError");
+const EventEmitter = require("events");
 
 /* exported OpenEVSE */
 class OpenEVSE extends EventEmitter
@@ -171,7 +17,7 @@ class OpenEVSE extends EventEmitter
   constructor (driver)
   {
     super();
-    this._version = "0.1";
+    this._version = require('./package.json').version;
     this._driver = driver;
 
     this.STATE_INVALID =                -1;
@@ -971,10 +817,13 @@ exports.connect = function(endpoint)
 {
   var driver;
   if("simulator" === endpoint) {
+    const OpenEVSEDriverSimulator = require("OpenEVSEDriverSimulator");
     driver = new OpenEVSEDriverSimulator();
   } else if(endpoint.startsWith("http:") || endpoint.startsWith("https:")) {
+    const OpenEVSEDriverHttp = require("OpenEVSEDriverHttp");
     driver = new OpenEVSEDriverHttp(endpoint);
   } else {
+    const OpenEVSEDriverSerial = require("OpenEVSEDriverSerial");
     driver = new OpenEVSEDriverSerial(endpoint);
   }
